backend/requirements.txt
# === Core ===
fastapi>=0.135.0
uvicorn[standard]>=0.30.1
gunicorn>=22.0.0
python-dotenv==1.0.1
pydantic==2.7.4
python-multipart==0.0.9

# === Database ===
sqlalchemy>=2.0.0
alembic>=1.18.0
psycopg2-binary>=2.9.0

# === Auth ===
bcrypt>=4.0.0
PyJWT>=2.8.0

# === AI / LLM ===
openai>=1.12.0

# === Payments ===
stripe==10.12.0

# === Google APIs ===
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# === Web / HTTP ===
requests==2.32.3
httpx>=0.27.0

# === Browser Automation ===
playwright==1.49.1

# === Utils ===
pyyaml==6.0.2
rich>=13.0.0

# === Redis ===
redis>=5.0.0

# === Monitoring ===
sentry-sdk[fastapi]>=2.0.0

# === Email ===
resend>=2.0.0

# === MCP (Model Context Protocol) ===
mcp[cli]>=1.0.0
